requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.0.0
arrow==1.2.3
asgiref==3.8.1
astropy==5.3.2
attrs==24.3.0
auto-py-to-exe==2.34.0
binaryornot==0.4.4
blockdiag==3.0.0
bottle==0.12.25
bottle-websocket==0.2.9
briefcase==0.3.14
build==0.10.0
buildozer==1.5.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cookiecutter==2.6.0
copier==8.3.0
customtkinter==5.1.2
cv==1.0.0
cvzone==1.6.1
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.6
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.19
duckdb==0.7.1
dunamai==1.19.0
Eel==0.16.0
email_validator==2.1.1
env==0.1.0
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.10.6
flatbuffers==24.12.23
flet==0.22.1
flet-core==0.22.1
flet-route==0.3.2
flet-runtime==0.22.1
flutter==0.1
fonttools==4.55.3
fpyo2apk==1.1.4
frozenlist==1.5.0
funcparserlib==1.0.1
funcy==2.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==2.0.2
groq==0.8.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.1
idna==3.10
imageio==2.26.0
imageio-ffmpeg==0.4.8
img2pdf==0.5.1
importlib-metadata==6.11.0
jax==0.5.0
jaxlib==0.5.0
jinja2-ansible-filters==1.3.2
joblib==1.3.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.7.0
Kivy-examples==2.0.0rc3
Kivy-Garden==0.1.5
kivy-garden.filebrowser==1.1.2
kivy-garden.graph==0.4.0
kiwisolver==1.4.8
lxml==5.1.0
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
moviepy==1.0.3
mss==9.0.1
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.3
packaging==24.2
Panda3D==1.10.14
panda3d-gltf==1.1.0
panda3d-simplepbr==0.11.2
pathspec==0.11.2
pefile==2023.2.7
pexpect==4.8.0
pikepdf==8.12.0
pillow==11.1.0
platformdirs==3.2.0
plumbum==1.8.2
proglog==0.1.10
prompt-toolkit==3.0.36
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyerfa==2.0.1.5
pygame==2.3.0
pyglet==1.5.16
pyinsane==1.4.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pylance==0.4.4
pymongo==4.7.0
Pympler==1.1
pyparsing==3.2.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
pyproject_hooks==1.0.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
python-slugify==8.0.1
python-vlc==3.0.18121
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
repath==0.9.0
reportlab==4.1.0
requests==2.32.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.15.1
screeninfo==0.8.1
sentencepiece==0.2.0
sh==2.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.0
sounddevice==0.5.1
sqlparse==0.4.3
starlette==0.37.2
text-unidecode==1.3
threadpoolctl==3.2.0
tkVideo==0.1
tomli_w==1.0.0
tqdm==4.67.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
ursina==6.1.2
uvicorn==0.29.0
validators==0.34.0
vboxapi==1.0
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
watchfiles==0.21.0
wcwidth==0.2.8
webcolors==1.13
websocket-client==1.6.3
websockets==11.0.3
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.16.0
yarl==1.18.3
zipp==3.21.0
zope.event==4.6
zope.interface==6.0
